EZP-23471: removed StorageRepositoryProvider
