+++ eZ/Bundle/EzPublishCoreBundle/Resources/config/storage_engines.yml
@@ -1,16 +1,9 @@
 parameters:
     ezpublish.persistence.connection.factory.class: eZ\Bundle\EzPublishCoreBundle\ApiLoader\StorageConnectionFactory
-    ezpublish.api.storage_repository_provider.class: eZ\Bundle\EzPublishCoreBundle\ApiLoader\StorageRepositoryProvider
     ezpublish.api.repository_configuration_provider.class: eZ\Bundle\EzPublishCoreBundle\ApiLoader\RepositoryConfigurationProvider
     ezpublish.persistence.connection.class: Doctrine\DBAL\Driver\Connection
 
 services:
-    ezpublish.api.storage_repository_provider:
-        class: %ezpublish.api.storage_repository_provider.class%
-        arguments:
-            - @ezpublish.config.resolver
-            - %ezpublish.repositories%
-
     ezpublish.api.repository_configuration_provider:
         class: %ezpublish.api.repository_configuration_provider.class%
         arguments:
